create advocates-table + search components
